Set up AutoComplete
I made it so that as a user types into the input field, the list is filtered down to Pokemon that starts with that query.

To emphasize the matching query, the query is wrapped in a <strong> tag.

I also set up a technique to capitalize the pokemon's name becuase it was lowercase by default.